pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dactyl_generation"
version = "0.0.1"
authors = [
    { name="Shantanu Thorat"},
]
description = "LLM helper package to generate examples."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "anthropic",
    "litellm",
    "mistralai",
    "numpy==1.26.4",
    "openai",
    "pandas",
    "protobuf",
    "python-dotenv",
    "Requests",
    "tqdm",
    "typing_extensions"
]
license = "MIT"
license-files = ["LICEN[CS]E*"]


